+++ blocks/header/header.js
@@ -1,6 +1,6 @@
 import { getMetadata } from '../../scripts/aem.js';
+import { loadFragment } from '../../scripts/scripts.js';
 import { appendXF } from '../experience-fragment/experience-fragment.js';
-import { loadFragment } from '../fragment/fragment.js';
 
 // media query match that indicates mobile/tablet width
 export const isDesktop = window.matchMedia('(min-width: 900px)');
