+++ scripts/scripts.js
@@ -12,23 +12,8 @@ import {
   loadSection,
   loadSections,
   loadCSS,
-  buildBlock,
   fetchPlaceholders,
 } from './aem.js';
-
-export async function loadErrorPage(main) {
-  if (window.errorCode === '404') {
-    const placeholder = await fetchPlaceholders();
-    const fragmentPath = placeholder.notFoundPage;
-    const fragmentLink = document.createElement('a');
-    fragmentLink.href = fragmentPath;
-    fragmentLink.textContent = fragmentPath;
-    const fragment = buildBlock('fragment', [[fragmentLink]]);
-
-    const section = main.querySelector('.section');
-    if (section) section.replaceChildren(fragment);
-  }
-}
 
 /**
  * Moves all the attributes from a given elmenet to another given element.
@@ -102,6 +87,51 @@ export function decorateMain(main) {
   decorateSections(main);
   decorateBlocks(main);
   buildAutoBlocks(main);
+}
+
+/**
+ * Loads a fragment.
+ * @param {string} path The path to the fragment
+ * @returns {HTMLElement} The root element of the fragment
+ */
+export async function loadFragment(path) {
+  if (path && path.startsWith('/')) {
+    // eslint-disable-next-line no-param-reassign
+    path = path.replace(/(\.plain)?\.html/, '');
+    const resp = await fetch(`${path}.plain.html`);
+    if (resp.ok) {
+      const main = document.createElement('main');
+      main.innerHTML = await resp.text();
+
+      // reset base path for media to fragment base
+      const resetAttributeBase = (tag, attr) => {
+        main.querySelectorAll(`${tag}[${attr}^="./media_"]`).forEach((elem) => {
+          elem[attr] = new URL(elem.getAttribute(attr), new URL(path, window.location)).href;
+        });
+      };
+      resetAttributeBase('img', 'src');
+      resetAttributeBase('source', 'srcset');
+
+      decorateMain(main);
+      await loadSections(main);
+      return main;
+    }
+  }
+  return null;
+}
+
+export async function loadErrorPage(main) {
+  if (window.errorCode === '404') {
+    const placeholder = await fetchPlaceholders();
+    // const fragmentPath = placeholder.notFoundPage;
+    // const fragmentLink = document.createElement('a');
+    // fragmentLink.href = window.location.origin + fragmentPath;
+    // fragmentLink.textContent = fragmentPath;
+    // const fragment = buildBlock('fragment', [[fragmentLink]]);
+    const fragment = await loadFragment(placeholder.notFoundPage);
+    const section = main.querySelector('.section');
+    if (section) section.replaceChildren(fragment);
+  }
 }
 
 /**
